#27: extract utils to separate file and create dedicated test folder
