+++ src/test/movie/movie.service.spec.ts
@@ -1,4 +1,4 @@
-import { MovieService } from './movie.service'
+import { MovieService } from '../../movie/movie.service'
 
 describe('MovieService', () => {
   let movieService: MovieService
